Add 64 bit integer support to StdCompiler in order to allow the last update timestamp to be saved as a 64 bit integer.

+++ src/C4Config.h
@@ -186,7 +186,7 @@ public:
 	bool LeagueAutoLogin;
 
 	bool AutomaticUpdate;
-	int32_t LastUpdateTime;
+	uint64_t LastUpdateTime;
 	int32_t AsyncMaxWait;
 
 public:
